+++ docs/CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [6.0.1] - 2019-09-05
 
+- add cobra metrics publisher + server unittest
+- add cobra client + server unittest
 - ws snake (cobra simple server) add basic support for unsubscription + subscribe send the proper subscription data + redis client subscription can be cancelled
 - IXCobraConnection / pdu handlers can crash if they receive json data which is not an object